Implement actor with goal-based AI behaviour
Actions could not be boxed before because they had type parameters
which depended on which actor was chosen.  Currently the only
Actor is actually the "monster" so just mering the two solves the
problem of being unable to be boxed.  Now, I can specify lots of
different actions and models to see how planning can be applied.