Enable tcod as basic backend for tui-rs
I want to use tui to drive the user interface of the demo,
implementing parts of the game as widgets in the tui framework.

Currently, however, tcod does not support unicode so I have to
match all of the symbols that the basic widgets put out in tcod to
the actual sprites that tcod supports and expects.  Eventually, the
goal would either be to have a unicode compliant backend to replace
tcod, or adapt the front-end of tui specifically to tcod.

Add bare minimum functionality to draw a frame, support for other
widgets coming.